skull emoji (Pulse Rifle arc fix)

+++ Projectiles/DraedonsArsenal/PulseRifleShot.cs
@@ -103,7 +103,7 @@ namespace CalamityMod.Projectiles.DraedonsArsenal
 
                     if (alreadyTargetedNPCType != Main.npc[i].whoAmI && Projectile.Center.ManhattanDistance(Main.npc[i].Center) < 600f)
                     {
-                        Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.Center, Projectile.SafeDirectionTo(Main.npc[i].Center) * 5f, Projectile.type, Projectile.damage * 0.4f, 0f, Projectile.owner, Main.npc[i].whoAmI, Projectile.ai[1] + 1f);
+                        Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.Center, Projectile.SafeDirectionTo(Main.npc[i].Center) * 5f, Projectile.type, (int)(Projectile.damage * 0.4f), 0f, Projectile.owner, Main.npc[i].whoAmI, Projectile.ai[1] + 1f);
                         break;
                     }
                 }
